+++ CHANGELOG.md
@@ -10,8 +10,9 @@
 - Moved `demos.md` to `DEMOS.md`.
 - Fixed the example JSON URL config in `TESTING.md`.
 - Fixed a horizontal scroll bug caused by overflow of the `{num} genes` subtitle in the Linnarsson demo.
-- Upgrade LayerController to be more general.
 - Fixed a regression caused by the updated bundling scripts minifying HTML assets and removing intentional spaces.
+- Upgrade LayerController to be more general.
+- In `src/app/app.js` and `src/demo/index.js`, separated rendering from validation. 
 
 ## [0.1.1](https://www.npmjs.com/package/vitessce/v/0.1.1) - 2020-05-05